added sidebar link enforcements
