+++ backend/internal/usersidebar/links.go
@@ -5,6 +5,25 @@ import (
 	"github.com/gtsteffaniak/filebrowser/backend/pkg/settings"
 )
 
+// prepareFrontendLink converts a stored sidebar link for API responses (source paths → display names).
+func prepareFrontendLink(link users.SidebarLink) (users.SidebarLink, bool) {
+	if !users.IsSourceSidebarCategory(link.Category) {
+		return link, true
+	}
+	source, ok := resolveSourceLink(link)
+	if !ok {
+		return link, false
+	}
+	if full, ok := settings.Config.Server.SourceMap[source.Path]; ok {
+		category := users.NormalizeSidebarLinkCategory(link.Category)
+		if full.Config.ResolvedRules.IndexingDisabled && category == string(users.SidebarLinkSource) {
+			link.Category = string(users.SidebarLinkSourceAlt)
+		}
+	}
+	link.SourceName = source.Name
+	return link, true
+}
+
 // FrontendLinks converts backend sidebar links to frontend-style links.
 func FrontendLinks(links []users.SidebarLink, showToolsInSidebar bool) []users.SidebarLink {
 	if !users.SourceConfigLoaded() {
@@ -17,17 +36,11 @@ func FrontendLinks(links []users.SidebarLink, showToolsInSidebar bool) []users.S
 			if link.SourceName == "" {
 				continue
 			}
-			source, ok := users.ResolveSourceKey(link.SourceName)
+			prepared, ok := prepareFrontendLink(link)
 			if !ok {
 				continue
 			}
-			if full, ok := settings.Config.Server.SourceMap[source.Path]; ok {
-				category := users.NormalizeSidebarLinkCategory(link.Category)
-				if full.Config.ResolvedRules.IndexingDisabled && category == string(users.SidebarLinkSource) {
-					link.Category = string(users.SidebarLinkSourceAlt)
-				}
-			}
-			link.SourceName = source.Name
+			link = prepared
 		} else if link.Category == "tool" && link.Target == "/tools" {
 			hasTools = true
 		}
