+++ frontend/src/components/prompts/SidebarLinks.vue
@@ -1,1163 +1,60 @@
 <template>
-  <div class="card-content sidebar-links-content">
-    <p v-if="!showAddForm && !yamlMode">{{ contextDescription }}</p>
-
-    <YamlEditorPanel
-      v-if="yamlMode"
-      v-model="yamlText"
-      @apply="applyYamlLinks"
-      @cancel="yamlMode = false"
-    />
-
-    <!-- Existing Links List - only show when not in edit/add mode -->
-    <div v-if="!showAddForm && !isSelectingPath && !yamlMode" class="links-list">
-      <h3>{{ $t('sidebar.currentLinks') }}</h3>
-      <div v-if="links.length === 0" class="empty-state">
-        <p>{{ $t('sidebar.noLinksYet') }}</p>
-      </div>
-      <div class="links-container">
-        <div v-for="(link, index) in links" :key="index" :ref="el => linkItemRefs[index] = el"
-          class="link-item input no-select" :class="{ 'dragging': draggingIndex === index }"
-          @dragover.prevent="handleDragOver($event, index)" @drop="handleDrop($event, index)">
-          <div draggable="true" @dragstart="handleDragStart($event, index)" @dragend="handleDragEnd"
-            class="link-drag-handle">
-            <i class="material-symbols">drag_indicator</i>
-          </div>
-          <div v-if="link.category === 'divider'" class="link-icon">
-            <i class="material-symbols">horizontal_rule</i>
-          </div>
-          <div v-else class="link-icon">
-            <i :class="getIconClass(link.icon)">{{ link.icon }}</i>
-          </div>
-          <div class="link-details">
-            <span class="link-name">{{ getLinkDisplayName(link) }}</span>
-            <span class="link-category">{{ getCategoryLabel(link.category) }}</span>
-          </div>
-          <button
-            type="button"
-            class="action"
-            @click="editLink(index)"
-            :disabled="isLinkEnforced(link)"
-            :aria-label="$t('general.edit')"
-            :title="isLinkEnforced(link) ? $t('profileSettings.enforcedByAdmin') : $t('general.edit')"
-          >
-            <i class="material-symbols">edit</i>
-          </button>
-          <button
-            type="button"
-            class="action"
-            @click="removeLink(index)"
-            :disabled="isLinkEnforced(link)"
-            :aria-label="$t('general.delete')"
-            :title="isLinkEnforced(link) ? $t('profileSettings.enforcedByAdmin') : $t('general.delete')"
-          >
-            <i class="material-symbols">delete</i>
-          </button>
-        </div>
-      </div>
-    </div>
-
-    <div v-if="!showAddForm && context === 'user' && !yamlMode" class="settings-items">
-      <ToggleSwitch class="item" :modelValue="showToolsInSidebar"
-        @update:modelValue="onShowToolsInSidebarChange"
-        :name="$t('profileSettings.showToolsInSidebar')"
-        :description="$t('profileSettings.showToolsInSidebarDescription')" />
-    </div>
-
-    <!-- Add New Link Section -->
-  <div v-if="!showAddForm && !yamlMode" class="add-link-section">
-      <button
-        type="button"
-        @click="showAddForm = true"
-        class="button button--flat button--blue add-link-button"
-      >
-        <i class="material-symbols">add</i>
-        {{ $t('sidebar.addNewLink') }}
-      </button>
-      <button
-        type="button"
-        class="button button--flat add-link-button"
-        @click="openYamlEditor"
-      >
-        {{ $t('settings.editAsYaml') }}
-      </button>
-    </div>
-
-    <!-- Add/Edit Link Form - replaces the list when active -->
-    <div v-else class="add-link-form">
-      <!-- Path Browser for Source/Share Links - shown when selecting path -->
-      <div v-if="isSelectingPath">
-        <file-list ref="fileList" :browse-source="isSourceCategory(newLink.category) ? newLink.sourceName : null"
-          :browse-share="newLink.category === 'share' ? getShareHash(newLink.target) : null"
-          @update:selected="updateSelectedPath"></file-list>
-      </div>
-
-      <!-- Form fields - hidden when selecting path -->
-      <div v-else>
-        <h3>{{ editingIndex !== null ? $t('sidebar.editLink') : $t('sidebar.addNewLink') }}</h3>
-
-        <!-- Link Type Selection -->
-        <p>{{ $t('sidebar.linkType') }}</p>
-        <ExpandDropdown
-          :model-value="linkTypeSelectValue"
-          :options="linkTypeOptions"
-          :default-placeholder-if-empty="$t('sidebar.selectLinkType')"
-          :aria-label="$t('sidebar.linkType')"
-          @update:model-value="onLinkTypeChange"
-        />
-
-        <!-- Source Selection (category is "source" or "source-minimal") -->
-        <div v-if="isSourceCategory(newLink.category)" class="form-group">
-          <p>{{ $t('sidebar.selectSource') }}</p>
-          <ExpandDropdown
-            v-model="newLink.sourceName"
-            :options="sidebarSourceOptions"
-            :default-placeholder-if-empty="$t('sidebar.chooseSource')"
-            :aria-label="$t('sidebar.selectSource')"
-            @update:model-value="handleSourceChange"
-          />
-
-          <!-- Custom Name for Source -->
-          <div class="form-group" v-if="newLink.sourceName">
-            <p>{{ $t('sidebar.linkName') }}</p>
-            <input aria-label="Link Name" v-model="newLink.name" type="text" class="input"
-              :placeholder="$t('sidebar.linkNamePlaceholder')" />
-          </div>
-
-          <!-- Path Selection for Source - clickable path display -->
-          <div v-if="newLink.sourceName" class="padding-top">
-            <div class="searchContext clickable button" @click="openPathBrowser('source')" aria-label="source-path">
-              {{ $t('general.path', { suffix: ':' }) }} {{ newLink.sourcePath || '/' }}
-            </div>
-          </div>
-
-          <!-- Source link options: Two toggles to control usage display type -->
-          <div v-if="newLink.sourceName" class="settings-items" style="margin-top: 0.5em;">
-            <!-- Show usage from indexed files toggle -->
-            <ToggleSwitch class="item"
-              :modelValue="showIndexedUsage"
-              @update:modelValue="updateUsageToggles('indexed', $event)"
-              :name="$t('sidebar.showIndexedUsage')"
-              :description="$t('sidebar.showIndexedUsageDescription')" />
-            <!-- Show disk/partition usage toggle -->
-            <ToggleSwitch class="item"
-              :modelValue="showDiskUsage"
-              @update:modelValue="updateUsageToggles('disk', $event)"
-              :name="$t('sidebar.showDiskUsage')"
-              :description="$t('sidebar.showDiskUsageDescription')" />
-            
-            <!-- Dropdown to choose which usage text to display (only shown in hybrid mode) -->
-            <div v-if="showIndexedUsage && showDiskUsage" class="form-group" style="margin-top: 0.5em;">
-              <p>{{ $t('sidebar.usageTextDisplay') }}</p>
-              <ExpandDropdown
-                :model-value="usageTextMode"
-                :options="usageTextModeOptions"
-                :aria-label="$t('sidebar.usageTextDisplay')"
-                @update:model-value="updateUsageTextMode"
-              />
-            </div>
-          </div>
-        </div>
-
-        <!-- Share Selection -->
-        <div v-if="newLink.category === 'share'" class="form-group">
-          <p>{{ $t('sidebar.selectShare') }}</p>
-          <ExpandDropdown
-            v-model="newLink.target"
-            :options="shareTargetOptions"
-            :default-placeholder-if-empty="$t('sidebar.chooseShare')"
-            :aria-label="$t('sidebar.selectShare')"
-            @update:model-value="handleShareChange"
-          />
-
-          <!-- Custom Name for Share -->
-          <div class="form-group" v-if="newLink.target">
-            <p>{{ $t('sidebar.linkName') }}</p>
-            <input aria-label="Link Name" v-model="newLink.name" type="text" class="input"
-              :placeholder="$t('sidebar.linkNamePlaceholder')" />
-          </div>
-
-          <!-- Path Selection for Share (subpath within the share) - clickable path display -->
-          <div v-if="newLink.target">
-            <div class="searchContext clickable button" @click="openPathBrowser('share')" aria-label="share-path">
-              {{ $t('general.path', { suffix: ':' }) }} {{ getShareSubpath(newLink.target) }}
-            </div>
-          </div>
-        </div>
-
-        <!-- Tool Selection - only available for user context, not shares -->
-        <div v-if="newLink.category === 'tool' && context === 'user'" class="form-group">
-          <p>{{ $t('sidebar.selectTool') }}</p>
-          <ExpandDropdown
-            v-model="newLink.target"
-            :options="toolTargetOptions"
-            :default-placeholder-if-empty="$t('sidebar.chooseTool')"
-            :aria-label="$t('sidebar.selectTool')"
-            @update:model-value="handleToolChange"
-          />
-
-          <!-- Custom Name for Tool -->
-          <div class="form-group" v-if="newLink.target">
-            <p>{{ $t('sidebar.linkName') }}</p>
-            <input aria-label="Link Name" v-model="newLink.name" type="text" class="input"
-              :placeholder="$t('sidebar.linkNamePlaceholder')" />
-          </div>
-        </div>
-
-        <!-- Share Info Link - special type for shares -->
-        <div v-if="newLink.category === 'shareInfo'" class="form-group">
-          <p>{{ $t('share.shareInfoDescription') }}</p>
-        </div>
-
-        <!-- Download Link - special type for shares -->
-        <div v-if="newLink.category === 'download'" class="form-group">
-          <p>{{ $t('share.downloadDescription') }}</p>
-        </div>
-
-        <!-- Divider - special visual separator -->
-        <div v-if="newLink.category === 'divider'" class="form-group">          
-          <p>{{ $t('sidebar.linkName') }}</p>
-          <input aria-label="Link Name" v-model="newLink.name" type="text" class="input"
-            :placeholder="$t('sidebar.dividerNamePlaceholder')" />
-        </div>
-
-        <!-- Custom Link Input -->
-        <div v-if="newLink.category === 'custom'" class="form-group">
-          <p>{{ $t('sidebar.linkName') }}</p>
-          <input aria-label="Link Name" v-model="newLink.name" type="text" class="input"
-            :placeholder="$t('sidebar.linkNamePlaceholder')" />
-
-          <p>{{ $t('sidebar.linkUrl') }}</p>
-          <input aria-label="Link Target" v-model="newLink.target" type="text" class="input"
-            :placeholder="$t('sidebar.linkUrlPlaceholder')" />
-        </div>
-
-        <!-- Icon Selection - Available for ALL link types except divider -->
-        <div v-if="newLink.category && newLink.category !== 'divider'" class="form-group">
-          <p>{{ $t('sidebar.linkIcon') }}</p>
-          <div class="icon-input-group">
-            <input v-model="newLink.icon" type="text" class="input icon-input"
-              :placeholder="$t('sidebar.linkIconPlaceholder')" />
-            <div class="icon-preview clickable" @click="openIconPicker" :title="$t('sidebar.browseIcons')">
-              <i v-if="newLink.icon" :class="getIconClass(newLink.icon)">{{ newLink.icon }}</i>
-              <i v-else class="material-symbols icon-preview-placeholder">interests</i>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div v-if="!yamlMode" class="card-actions">
-    <!-- When selecting a path -->
-    <template v-if="isSelectingPath">
-      <button
-        type="button"
-        @click="cancelPathSelection"
-        class="button button--flat button--grey"
-        :aria-label="$t('general.cancel')"
-        :title="$t('general.cancel')"
-      >
-        {{ $t("general.cancel") }}
-      </button>
-      <button
-        type="button"
-        @click="confirmPathSelection"
-        class="button button--flat button--blue"
-        :aria-label="$t('general.ok')"
-        :title="$t('general.ok')"
-      >
-        {{ $t("general.ok") }}
-      </button>
-    </template>
-
-    <!-- When in add/edit form mode -->
-    <template v-else-if="showAddForm">
-      <button
-        type="button"
-        @click="cancelAddLink"
-        class="button button--flat button--grey"
-        :aria-label="$t('general.cancel')"
-        :title="$t('general.cancel')"
-      >
-        {{ $t("general.cancel") }}
-      </button>
-      <button
-        type="button"
-        aria-label="Add Link"
-        @click="addLink"
-        class="button button--flat button--blue"
-        :disabled="!isNewLinkValid"
-        :title="editingIndex !== null ? $t('general.save') : $t('general.add')"
-      >
-        {{ editingIndex !== null ? $t('general.save') : $t('general.add') }}
-      </button>
-    </template>
-
-    <!-- When viewing the list -->
-    <template v-else>
-      <button
-        type="button"
-        aria-label="Save Links"
-        class="button button--flat button--blue"
-        @click="saveLinks"
-        :title="$t('general.save')"
-      >
-        {{ $t("general.save") }}
-      </button>
-    </template>
-  </div>
+  <SidebarLinksEditor
+    ref="editor"
+    :mode="context"
+    :share-data="shareData"
+    :show-prompt-actions="true"
+    @save="saveLinks"
+  />
 </template>
 
 <script>
-import { state, getters, mutations } from "@/store";
 import { notify } from "@/notify";
+import { mutations } from "@/store";
 import { shareApi } from "@/api";
-import { tools } from "@/utils/constants";
-import { getIconClass } from "@/utils/material-symbols";
-import { getObjectProperty } from '@/utils/object.js';
-import FileList from "../files/FileList.vue";
-import ToggleSwitch from "@/components/settings/ToggleSwitch.vue";
-import ExpandDropdown from "@/components/settings/ExpandDropdown.vue";
-import YamlEditorPanel from "@/components/prompts/YamlEditorPanel.vue";
 import { eventBus } from "@/store/eventBus";
-import { sidebarLinkKey } from "@/utils/sidebarLinkKeys.js";
-import yaml from "js-yaml";
+import SidebarLinksEditor from "@/components/sidebar/SidebarLinksEditor.vue";
 
 export default {
   name: "SidebarLinks",
   components: {
-    FileList,
-    ToggleSwitch,
-    ExpandDropdown,
-    YamlEditorPanel,
+    SidebarLinksEditor,
   },
   props: {
     context: {
       type: String,
-      default: "user", // 'user' or 'share'
+      default: "user",
     },
     shareData: {
       type: Object,
-      default: null, // Share object when context is 'share'
+      default: null,
     },
-  },
-  data() {
-    return {
-      links: [],
-      showAddForm: false,
-      newLink: {
-        name: "",
-        category: "",
-        target: "",
-        icon: "",
-        sourceName: "",
-        sourcePath: "",
-      },
-      draggingIndex: null,
-      dragOverIndex: null,
-      linkItemRefs: {},
-      originalLinks: null, // Store original order in case drag is cancelled
-      availableTools: [
-        { name: 'tools.title', path: '/tools', icon: 'build' }, // Main tools page
-        ...tools() // Individual tools
-      ],
-      availableShares: [],
-      editingIndex: null,
-      isSelectingPath: false,
-      tempSelectedPath: "",
-      tempSelectedSource: "",
-      showToolsInSidebar: true,
-      yamlMode: false,
-      yamlText: "",
-    };
-  },
-  computed: {
-    enforcedLinkKeys() {
-      const keys = state.user?.enforcedSidebarLinkKeys;
-      return Array.isArray(keys) ? new Set(keys) : new Set();
-    },
-    availableSources() {
-      return state.sources?.info || {};
-    },
-    contextDescription() {
-      return this.context === 'share'
-        ? this.$t('sidebar.customizeShareLinksDescription')
-        : this.$t('sidebar.customizeLinksDescription');
-    },
-    isNewLinkValid() {
-      if (!this.newLink.category) return false;
-
-      // Special link types that don't need additional validation
-      if (this.newLink.category === "shareInfo" || 
-          this.newLink.category === "download" || 
-          this.newLink.category === "divider") {
-        return true;
-      }
-
-      if (this.newLink.category === "custom") {
-        return this.newLink.name && this.newLink.target;
-      }
-
-      if (this.isSourceCategory(this.newLink.category)) {
-        return this.newLink.sourceName && this.newLink.name;
-      }
-
-      if (this.newLink.category === "share") {
-        return this.newLink.target && this.newLink.name;
-      }
-
-      return this.newLink.target && this.newLink.name;
-    },
-    showIndexedUsage() {
-      return this.newLink.category === 'source' || this.newLink.category === 'source-hybrid' || this.newLink.category === 'source-hybrid-2';
-    },
-    showDiskUsage() {
-      return this.newLink.category === 'source-alt' || this.newLink.category === 'source-hybrid' || this.newLink.category === 'source-hybrid-2';
-    },
-    usageTextMode() {
-      if (this.newLink.category === 'source-hybrid-2') {
-        return 'disk';
-      }
-      return 'indexed';
-    },
-    /** Single "Source" row in the link-type dropdown; actual variant stays in newLink.category (toggles). */
-    linkTypeSelectValue() {
-      if (this.isSourceCategory(this.newLink.category)) {
-        return 'source';
-      }
-      return this.newLink.category;
-    },
-    /** Share link type requires manage-share permission; avoids listing shares when forbidden */
-    canListShares() {
-      return this.context === "user" && getters.globalPermissions().share;
-    },
-    linkTypeOptions() {
-      const options = [];
-      if (this.context === "user") {
-        options.push({ value: "source", label: this.$t("general.source") });
-      }
-      if (this.canListShares) {
-        options.push({ value: "share", label: this.$t("general.share") });
-      }
-      if (this.context === "user") {
-        options.push({ value: "tool", label: this.$t("general.tool") });
-      }
-      options.push(
-        { value: "custom", label: this.$t("sidebar.customLink") },
-        { value: "divider", label: this.$t("general.divider") },
-      );
-      if (this.context === "share") {
-        options.push(
-          { value: "shareInfo", label: this.$t("share.shareInfo") },
-          { value: "download", label: this.$t("general.download") },
-        );
-      }
-      return options;
-    },
-    sidebarSourceOptions() {
-      return Object.keys(this.availableSources).map((name) => ({
-        value: name,
-        label: name,
-      }));
-    },
-    shareTargetOptions() {
-      return this.availableShares.map((share) => ({
-        value: `/public/share/${share.hash}`,
-        label: `${share.hash} ${this.$t("general.of")} ${share.path}`,
-      }));
-    },
-    toolTargetOptions() {
-      return this.availableTools.map((tool) => ({
-        value: tool.path,
-        label: this.$t(tool.name),
-      }));
-    },
-    usageTextModeOptions() {
-      return [
-        { value: "indexed", label: this.$t("sidebar.usageTextIndexed") },
-        { value: "disk", label: this.$t("sidebar.usageTextDisk") },
-      ];
-    },
-  },
-  async mounted() {
-    // Initialize with existing sidebar links based on context
-    if (this.context === 'share' && this.shareData?.sidebarLinks) {
-      this.links = [...this.shareData.sidebarLinks];
-    } else if (this.context === 'user' && state.user?.sidebarLinks && state.user?.sidebarLinks.length > 0) {
-      this.links = [...state.user.sidebarLinks];
-    } else if (this.context === 'user') {
-      // Generate default links from sources for user context
-      this.links = this.getDefaultLinks();
-    }
-
-    if (this.context === 'user') {
-      if (typeof state.user?.showToolsInSidebar === 'boolean') {
-        this.showToolsInSidebar = state.user?.showToolsInSidebar;
-      } else {
-        this.showToolsInSidebar = true;
-      }
-      this.syncMainToolsHubLinkRow(this.showToolsInSidebar);
-    }
-
-    // Share list loads lazily when adding/editing a share link (requires permission).
   },
   methods: {
-    /** Main Tools hub uses category "tool" and target "/tools"; keep in sync when the toggle changes. */
-    isMainToolsHubLink(link) {
-      return link?.category === "tool" && link?.target === "/tools";
-    },
-    syncMainToolsHubLinkRow(showHub) {
-      if (this.context !== "user") return;
-      if (!showHub) {
-        this.links = this.links.filter((link) => !this.isMainToolsHubLink(link));
-        return;
-      }
-      if (this.links.some((link) => this.isMainToolsHubLink(link))) {
-        return;
-      }
-      const mainEntry = this.availableTools.find((t) => t.path === "/tools");
-      const name = mainEntry ? this.$t(mainEntry.name) : this.$t("tools.title");
-      const icon = mainEntry ? mainEntry.icon : "build";
-      this.links.push({
-        name,
-        category: "tool",
-        target: "/tools",
-        icon,
-      });
-    },
-    onShowToolsInSidebarChange(value) {
-      this.showToolsInSidebar = value;
-      this.syncMainToolsHubLinkRow(value);
-    },
-    isLinkEnforced(link) {
-      if (this.context !== "user" || getters.isAdmin()) {
-        return false;
-      }
-      return this.enforcedLinkKeys.has(sidebarLinkKey(link));
-    },
-    openYamlEditor() {
-      this.yamlText = yaml.dump(this.links, { lineWidth: 120, noRefs: true });
-      this.yamlMode = true;
-    },
-    applyYamlLinks(text) {
+    async saveLinks({ links, showToolsInSidebar }) {
       try {
-        const parsed = yaml.load(text);
-        if (!Array.isArray(parsed)) {
-          throw new Error("expected array");
-        }
-        this.links = parsed.map((link) => ({ ...link }));
-        this.yamlMode = false;
-        void this.saveLinks();
-      } catch (_e) {
-        notify.showError(this.$t("settings.sidebarLinkDefaultsYamlInvalid"));
-      }
-    },
-    getIconClass,
-    getShareHash(target) {
-      // Extract hash from /public/share/<hash> or /public/share/<hash>/path
-      if (!target) return '';
-      const parts = target.split('/');
-      // parts: ['', 'public', 'share', '<hash>', ...subpath]
-      if (parts.length >= 4 && parts[1] === 'public' && parts[2] === 'share') {
-        return parts[3];
-      }
-      return '';
-    },
-    getShareSubpath(target) {
-      // Extract subpath from /public/share/<hash>/subpath
-      if (!target) return '/';
-      const parts = target.split('/');
-      // parts: ['', 'public', 'share', '<hash>', ...subpath]
-      if (parts.length >= 4 && parts[1] === 'public' && parts[2] === 'share') {
-        return parts.length > 4 ? `/${parts.slice(4).join('/')}` : '/';
-      }
-      return '/';
-    },
-    openIconPicker() {
-      mutations.showPrompt({
-        name: "IconPicker",
-        props: {
-          onSelect: this.handleIconSelect,
-        },
-      });
-    },
-    handleIconSelect(iconName) {
-      this.newLink.icon = iconName;
-    },
-    async loadAvailableShares() {
-      if (!this.canListShares) {
-        this.availableShares = [];
-        return;
-      }
-      try {
-        this.availableShares = await shareApi.list();
-      } catch (error) {
-        console.error("Failed to load shares:", error);
-        this.availableShares = [];
-      }
-    },
-    getDefaultLinks() {
-      // Generate default links from sources
-      const defaultLinks = [];
-
-      if (this.availableSources) {
-        Object.keys(this.availableSources).forEach(sourceName => {
-          defaultLinks.push({
-            name: sourceName,
-            category: 'source',
-            target: '/', // Relative path to source root
-            icon: '', // No icon by default - will show animated status indicator
-            sourceName: sourceName,
-          });
-        });
-      }
-
-      return defaultLinks;
-    },
-    isSourceCategory(category) {
-      return category === 'source' || category === 'source-minimal' || category === 'source-alt' || category === 'source-hybrid' || category === 'source-hybrid-2';
-    },
-    updateUsageToggles(toggleType, value) {
-      // Determine the new category based on toggle states
-      // indexed=true, disk=false  -> 'source'
-      // indexed=false, disk=true  -> 'source-alt'
-      // indexed=true, disk=true   -> 'source-hybrid' or 'source-hybrid-2' (depends on usageTextMode)
-      // indexed=false, disk=false -> 'source-minimal'
-      
-      const indexed = toggleType === 'indexed' ? value : this.showIndexedUsage;
-      const disk = toggleType === 'disk' ? value : this.showDiskUsage;
-      
-      if (indexed && disk) {
-        // Preserve the hybrid mode variant if it was already set
-        if (this.newLink.category === 'source-hybrid-2') {
-          this.newLink.category = 'source-hybrid-2';
-        } else {
-          this.newLink.category = 'source-hybrid';
-        }
-      } else if (indexed && !disk) {
-        this.newLink.category = 'source';
-      } else if (!indexed && disk) {
-        this.newLink.category = 'source-alt';
-      } else {
-        this.newLink.category = 'source-minimal';
-      }
-    },
-    updateUsageTextMode(mode) {
-      if (mode === "disk") {
-        this.newLink.category = 'source-hybrid-2';
-      } else {
-        this.newLink.category = 'source-hybrid';
-      }
-    },
-    getCategoryLabel(category) {
-      switch (category) {
-        case 'source':
-        case 'source-minimal':
-        case 'source-alt':
-        case 'source-hybrid':
-        case 'source-hybrid-2':
-          return this.$t('general.source');
-        case 'tool':
-          return this.$t('general.tool');
-        case 'custom':
-          return this.$t('sidebar.customLink');
-        case 'share':
-          return this.$t('general.share');
-        case 'shareInfo':
-          return this.$t('share.shareInfo');
-        case 'download':
-          return this.$t('general.download');
-        case 'divider':
-          return this.$t('general.divider');
-        default:
-          return category;
-      }
-    },
-    getLinkDisplayName(link) {
-      // For dividers, show the name or a default
-      if (link.category === 'divider') {
-        return link.name || this.$t('general.divider');
-      }
-      
-      // Check if the name looks like a translation key that needs translating
-      if (link.category === 'shareInfo' && link.name === 'share.shareInfo') {
-        return this.$t('share.shareInfo');
-      }
-      if (link.category === 'download' && link.name === 'general.download') {
-        return this.$t('general.download');
-      }
-      // Check if it's a general translation key pattern
-      if (typeof link.name === 'string' && link.name.includes('.') && link.name.split('.').length === 2) {
-        // Try to translate, if it fails, return original
-        try {
-          const translated = this.$t(link.name);
-          // If translation returns the same key, it means it doesn't exist, return original
-          return translated !== link.name ? translated : link.name;
-        } catch (_e) {
-          return link.name;
-        }
-      }
-      return link.name;
-    },
-    resetNewLinkFormForTypeSwitch() {
-      this.newLink.name = "";
-      this.newLink.target = "";
-      this.newLink.icon = "";
-      this.newLink.sourceName = "";
-      this.newLink.sourcePath = "";
-    },
-    onLinkTypeChange(value) {
-      if (value === "source") {
-        if (!this.isSourceCategory(this.newLink.category)) {
-          this.resetNewLinkFormForTypeSwitch();
-          this.newLink.category = "source";
-        }
-        return;
-      }
-      if (value !== this.newLink.category || this.isSourceCategory(this.newLink.category)) {
-        this.resetNewLinkFormForTypeSwitch();
-        this.newLink.category = value;
-        if (value === "shareInfo") {
-          this.newLink.name = this.$t("share.shareInfo");
-          this.newLink.icon = "qr_code";
-        } else if (value === "download") {
-          this.newLink.name = this.$t("general.download");
-          this.newLink.icon = "download";
-        }
-      }
-      if (value === "share") {
-        void this.loadAvailableShares();
-      }
-    },
-    handleSourceChange() {
-      if (this.newLink.sourceName) {
-        // Only set default name if user hasn't entered one yet
-        if (!this.newLink.name) {
-          this.newLink.name = this.newLink.sourceName;
-        }
-        // No icon by default - will show animated status indicator
-        if (!this.newLink.icon) {
-          this.newLink.icon = "";
-        }
-        this.newLink.sourcePath = "/";
-      }
-    },
-    handleShareChange() {
-      if (this.newLink.target) {
-        const hash = this.getShareHash(this.newLink.target);
-        const share = this.availableShares.find(s => s.hash === hash);
-        if (share) {
-          // Only set default name if user hasn't entered one yet
-          if (!this.newLink.name) {
-            this.newLink.name = `Share: ${share.hash}`;
-          }
-          // Suggest default icon if not set
-          if (!this.newLink.icon) {
-            this.newLink.icon = "share";
-          }
-        }
-      }
-    },
-    handleToolChange() {
-      const tool = this.availableTools.find(t => t.path === this.newLink.target);
-      if (tool) {
-        // Only set default name and icon if user hasn't entered them yet
-        if (!this.newLink.name) {
-          this.newLink.name = this.$t(tool.name);
-        }
-        if (!this.newLink.icon) {
-          this.newLink.icon = tool.icon;
-        }
-      }
-    },
-    openPathBrowser(type) {
-      // Show file list for path browsing
-      this.isSelectingPath = true;
-      if (type === 'source') {
-        this.tempSelectedPath = this.newLink.sourcePath || '/';
-        this.tempSelectedSource = this.newLink.sourceName;
-      } else if (type === 'share') {
-        this.tempSelectedPath = this.getShareSubpath(this.newLink.target);
-        this.tempSelectedSource = '';
-      }
-    },
-    updateSelectedPath(pathOrData) {
-      // Handle both old format (string) and new format (object with path and source)
-      if (typeof pathOrData === 'string') {
-        this.tempSelectedPath = pathOrData;
-      } else if (pathOrData?.path) {
-        this.tempSelectedPath = pathOrData.path;
-        this.tempSelectedSource = pathOrData.source;
-      }
-    },
-    confirmPathSelection() {
-      // Apply the selected path to the link based on category
-      if (this.isSourceCategory(this.newLink.category)) {
-        this.newLink.sourcePath = this.tempSelectedPath;
-      } else if (this.newLink.category === 'share') {
-        // Update target with new subpath
-        const hash = this.getShareHash(this.newLink.target);
-        const subpath = this.tempSelectedPath === '/' ? '' : this.tempSelectedPath;
-        this.newLink.target = `/public/share/${hash}${subpath}`;
-      }
-      this.isSelectingPath = false;
-    },
-    cancelPathSelection() {
-      // Cancel path selection and return to form
-      this.isSelectingPath = false;
-      this.tempSelectedPath = "";
-      this.tempSelectedSource = "";
-    },
-    editLink(index) {
-      const link = this.links.at(index);
-      if (!link || this.isLinkEnforced(link)) return;
-      this.editingIndex = index;
-      this.showAddForm = true;
-
-      // Populate form with existing link data (category stays source or source-minimal)
-      const isSource = this.isSourceCategory(link.category);
-      this.newLink = {
-        name: link.name,
-        category: link.category,
-        target: isSource ? "" : (link.target || ""),
-        icon: link.icon || "",
-        sourceName: link.sourceName || "",
-        sourcePath: isSource ? (link.target || "/") : "/",
-      };
-      if (link.category === "share") {
-        void this.loadAvailableShares();
-      }
-    },
-    addLink() {
-      if (!this.isNewLinkValid) return;
-
-      // Build the link object based on category
-      // Always include: name, category, target, icon, and conditionally sourceName/shareHash
-      const linkData = {
-        name: this.newLink.name,
-        category: this.newLink.category,
-        icon: this.newLink.icon,
-        target: "",
-      };
-
-      // Process target and additional fields based on category
-      if (this.newLink.category === "shareInfo") {
-        // ShareInfo is a special action link
-        linkData.target = "#";
-      } else if (this.newLink.category === "download") {
-        // Download is a special action link
-        linkData.target = "#";
-      } else if (this.newLink.category === "divider") {
-        // Divider is a visual separator with no action
-        linkData.target = "#";
-        linkData.name = this.newLink.name || ""; // Keep the custom name if provided
-        linkData.icon = ""; // Dividers don't need an icon
-      } else if (this.newLink.category === "custom") {
-        linkData.target = this.processCustomUrl(this.newLink.target);
-      } else if (this.isSourceCategory(this.newLink.category)) {
-        // For sources: target is relative path, sourceName identifies the source; category is "source" or "source-minimal"
-        linkData.category = this.newLink.category;
-        linkData.target = this.newLink.sourcePath || '/';
-        linkData.sourceName = this.newLink.sourceName;
-      } else if (this.newLink.category === "share") {
-        // For shares: target is already the full path /public/share/<hash>/<subpath>
-        linkData.target = this.newLink.target;
-      } else if (this.newLink.category === "tool") {
-        linkData.target = this.newLink.target;
-      }
-
-      if (this.editingIndex !== null) {
-        // Update existing link
-        this.links[this.editingIndex] = linkData;
-      } else {
-        // Add new link
-        this.links.push(linkData);
-      }
-
-      // Close the form and return to list view
-      this.cancelAddLink();
-    },
-    processCustomUrl(url) {      
-      // Check if it's an external URL (case insensitive)
-      const lowerUrl = url.toLowerCase();
-      if (lowerUrl.startsWith('http://') || lowerUrl.startsWith('https://')) {
-        // Leave external URLs as-is
-        return url;
-      }
-      // For internal paths, normalize by adding a starting slash if missing
-      if (!url.startsWith('/')) {
-        return `/${url}`; // assume relative path
-      }
-      return url;
-    },
-    removeLink(index) {
-      this.links.splice(index, 1);
-    },
-    cancelAddLink() {
-      this.showAddForm = false;
-      this.editingIndex = null;
-      this.newLink = {
-        name: "",
-        category: "",
-        target: "",
-        icon: "",
-        sourceName: "",
-        sourcePath: "",
-      };
-    },
-    handleDragStart(event, index) {
-      this.draggingIndex = index;
-      this.dragOverIndex = null;
-
-      // Store original order in case drag is cancelled
-      this.originalLinks = [...this.links];
-
-      event.dataTransfer.effectAllowed = "move";
-      event.dataTransfer.setData("text/html", event.target);
-
-      // Set the entire link item as the drag image
-      const linkItem = getObjectProperty(this.linkItemRefs, index);
-      if (linkItem) {
-        // Create a clone for the drag image to avoid affecting the original
-        const clone = linkItem.cloneNode(true);
-        clone.style.position = 'fixed';
-        clone.style.top = '-9999px';
-        clone.style.left = '-9999px';
-
-        // Set the clone width to match the original element's width
-        const originalWidth = linkItem.offsetWidth;
-        clone.style.width = `${originalWidth}px`;
-        linkItem.parentNode.insertBefore(clone, linkItem.nextSibling);
-
-        // Set it as the drag image
-        event.dataTransfer.setDragImage(clone, event.offsetX, event.offsetY);
-
-        // Clean up the clone after a brief delay
-        setTimeout(() => {
-          clone.remove();
-        }, 0);
-      }
-    },
-    handleDragOver(_event, index) {
-      if (this.draggingIndex === null || this.draggingIndex === index) return;
-
-      // Only reorder if we're hovering over a different item
-      if (this.dragOverIndex !== index) {
-        this.dragOverIndex = index;
-
-        // Live reorder: move the dragged item to the new position
-        const newLinks = [...this.links];
-        const draggedLink = newLinks[this.draggingIndex];
-
-        // Remove from current position
-        newLinks.splice(this.draggingIndex, 1);
-
-        // Insert at hover position
-        newLinks.splice(index, 0, draggedLink);
-
-        // Update the array and dragging index
-        this.links = newLinks;
-        this.draggingIndex = index; // Update to new position
-      }
-    },
-    handleDrop(event) {
-      event.preventDefault();
-
-      // The array is already in the correct order from handleDragOver
-      // Just clean up the drag state
-      this.draggingIndex = null;
-      this.dragOverIndex = null;
-      this.originalLinks = null; // Clear the backup
-    },
-    handleDragEnd() {
-      // If drag was cancelled (no drop event), restore original order
-      if (this.originalLinks !== null) {
-        this.links = this.originalLinks;
-        this.originalLinks = null;
-      }
-
-      this.draggingIndex = null;
-      this.dragOverIndex = null;
-    },
-    async saveLinks() {
-      try {
-        if (this.context === 'share') {
-          // Save to share
+        if (this.context === "share") {
           const payload = {
             hash: this.shareData.hash,
-            sidebarLinks: this.links,
+            sidebarLinks: links,
           };
 
           await shareApi.create(payload);
 
-          // Notify Share component of the updated links via eventBus
-          eventBus.emit('shareSidebarLinksUpdated', {
+          eventBus.emit("shareSidebarLinksUpdated", {
             hash: this.shareData.hash,
-            sidebarLinks: this.links,
+            sidebarLinks: links,
           });
-
         } else {
-          // Save to user
           await mutations.updateCurrentUser({
-            sidebarLinks: [...this.links],
-            showToolsInSidebar: this.showToolsInSidebar,
+            sidebarLinks: [...links],
+            showToolsInSidebar,
           });
-
-          notify.showSuccessToast(this.$t("sidebar.linksUpdatedSuccess"));
         }
-
-        // Close only this prompt, returning to the previous one (if any)
-        mutations.closeTopPrompt();
-      } catch (_e) {
-        notify.showError(this.$t("sidebar.linksUpdateFailed"));
+      } catch (e) {
+        notify.showError(e);
       }
     },
   },
 };
 </script>
-
-<style scoped>
-
-.settings-items {
-  margin-top: 0.5em;
-  margin-bottom: 0.5em;
-}
-
-.padding-top {
-  margin-top: 0.5em;
-}
-
-.links-list h3,
-.add-link-form h3 {
-  margin-bottom: 0.5em;
-  font-size: 1em;
-  font-weight: 600;
-}
-
-.empty-state {
-  padding: 2em 1em;
-  text-align: center;
-  color: var(--textSecondary);
-  font-style: italic;
-}
-
-.add-link-form h3 {
-  margin-top: 0;
-  margin-bottom: 0.75em;
-  font-size: 0.95em;
-}
-
-.links-container {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5em;
-}
-
-/* Link item styles */
-.link-item {
-  display: flex;
-  align-items: center;
-  gap: 0.5em;
-  background: var(--surfaceSecondary);
-  transition: all 0.2s ease;
-}
-
-.link-item.dragging {
-  opacity: 0.5;
-  border-color: var(--primaryColor);
-  background: var(--surfaceTertiary);
-}
-
-.link-drag-handle {
-  color: var(--textSecondary);
-  cursor: grab;
-}
-
-.link-drag-handle:active {
-  cursor: grabbing;
-}
-
-.link-icon {
-  color: var(--primaryColor);
-}
-
-.link-details {
-  display: flex;
-  flex-direction: column;
-  gap: 0.25em;
-  flex-grow: 1;
-  width: 100%;
-}
-
-.link-name {
-  font-weight: 500;
-}
-
-.link-category {
-  font-size: 0.85em;
-  color: var(--textSecondary);
-}
-
-.add-link-section {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5em;
-}
-
-.add-link-button {
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5em;
-}
-
-.add-link-form {
-  padding: 0;
-  margin-top: 0;
-}
-
-.form-group p {
-  margin: 0.5em
-}
-
-.form-group p:first-of-type,
-.add-link-form>p:first-of-type {
-  margin-top: 0;
-}
-
-/* Icon preview styles */
-.icon-input-group {
-  display: flex;
-  align-items: center;
-  gap: 0.5em;
-}
-
-.icon-input {
-  flex: 1;
-}
-
-.icon-preview {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 3em;
-  height: 3em;
-  border: 1px solid var(--borderColor);
-  border-radius: 1em;
-  background: var(--surfaceSecondary);
-  color: var(--primaryColor);
-}
-
-.icon-preview .material-symbols,
-.icon-preview .material-symbols-outlined {
-  font-size: 2em;
-}
-
-.icon-preview-placeholder {
-  color: var(--textSecondary);
-  opacity: 0.6;
-}
-
-</style>
