+++ frontend/src/i18n/ar.json
@@ -298,7 +298,8 @@
     "accountDefaults": "التخلف عن سداد الالتزامات المتعلقة بالحساب",
     "disableUserSettings": "تعطيل صفحة إعدادات المستخدم",
     "userDefaultFieldLockedFromConfig": "يتم تعيين هذا الإعداد الافتراضي في ملف التكوين ولا يمكن تغييره هنا.",
-    "sourceAccessDefaultsHelp": "يتم تطبيق أذونات الملفات الافتراضية على المصادر والمستخدمين الجدد. ويؤدي خيار «Enforce» إلى تقييد القيمة لجميع المستخدمين غير الإداريين."
+    "sourceAccessDefaultsHelp": "يتم تطبيق أذونات الملفات الافتراضية على المصادر والمستخدمين الجدد. ويؤدي خيار «Enforce» إلى تقييد القيمة لجميع المستخدمين غير الإداريين.",
+    "editAsYaml": "تحرير بتنسيق YAML"
   },
   "sidebar": {
     "updateIsAvailable": "يتوفر تحديث!",
@@ -319,10 +320,7 @@
     "linkUrl": "رابط الرابط URL",
     "linkUrlPlaceholder": "أدخل عنوان URL أو المسار (على سبيل المثال، /ملفات/مسار أو https://example.com)",
     "linkIcon": "أيقونة الرابط",
-    "linksUpdatedSuccess": "تم تحديث روابط الشريط الجانبي بنجاح!",
-    "linksUpdateFailed": "فشل في تحديث روابط الشريط الجانبي",
     "sourceNotAccessible": "لم يعد هذا المصدر متاحاً لك",
-    "noLinksYet": "لم يتم تكوين أي روابط حتى الآن. أضف الرابط الأول أدناه.",
     "customizeShareLinksDescription": "قم بتخصيص الروابط المعروضة في الشريط الجانبي لهذه المشاركة. السحب والإفلات لإعادة الترتيب.",
     "editLink": "تحرير الرابط",
     "selectShare": "حدد المشاركة",
@@ -339,7 +337,12 @@
     "usageTextDisk": "الاستخدام كما تم الإبلاغ عنه بواسطة القرص/القسم",
     "dividerNamePlaceholder": "نص المقسم (اتركه فارغًا لإظهار سطر)",
     "showIndexedUsage": "عرض الاستخدام حسب الفهرس",
-    "showDiskUsage": "عرض الاستخدام وفقًا لما أورده القرص/القسم"
+    "showDiskUsage": "عرض الاستخدام وفقًا لما أورده القرص/القسم",
+    "sidebarLinkDefaults": "الإعدادات الافتراضية لروابط الشريط الجانبي",
+    "sidebarLinkDefaultsDescription": "تكوين روابط الشريط الجانبي الافتراضية والإلزامية لجميع المستخدمين.",
+    "sidebarLinkDefaultsHelp": "هذه هي الروابط الافتراضية التي ستُضاف لجميع المستخدمين الجدد. إذا قمت بتفعيل رابط ما، فسيتم إضافته لجميع المستخدمين ولن يتمكن المستخدمون غير الإداريين من إزالته.",
+    "sidebarLinkDefaultSourceHelp": "ينطبق هذا فقط عندما يكون لدى المستخدم حق الوصول إلى هذا المصدر.",
+    "sidebarLinkDefaultsYamlInvalid": "صيغة YAML غير صالحة. تحقق من التنسيق وحاول مرة أخرى."
   },
   "share": {
     "notice": "ملاحظة: عندما يتم إنشاء مشاركة، تصبح أكثر انكشافًا للجمهور ولها قواعد مختلفة للوصول إليها. إذا لم يتم تهيئة مشاركتك بشكل آمن، فيمكن لأي شخص الوصول إليها.",
@@ -620,7 +623,8 @@
     "limit": "الحد الأقصى{suffix}",
     "remove": "إزالة{suffix}",
     "saved": "تم الحفظ{suffix}",
-    "removed": "تم حذفه{suffix}"
+    "removed": "تم حذفه{suffix}",
+    "back": "العودة{suffix}"
   },
   "access": {
     "rules": "قواعد الوصول",
@@ -918,7 +922,6 @@
       "browseFullLibrary": "تصفح مكتبة الأيقونات الكاملة"
     },
     "toolNotFound": "الأداة غير موجودة",
-    "backToTools": "العودة إلى الأدوات",
     "fileWatcher": {
       "name": "مراقب الملفات",
       "description": "مشاهدة وتعقب الملفات في الوقت الفعلي مع مراقبة زمن الاستجابة",
