+++ frontend/src/i18n/bg.json
@@ -167,7 +167,8 @@
     "limit": "Лимит{suffix}",
     "remove": "Премахни{suffix}",
     "saved": "Запазено{suffix}",
-    "removed": "Премахнато{suffix}"
+    "removed": "Премахнато{suffix}",
+    "back": "Назад{suffix}"
   },
   "buttons": {
     "copySuccess": "Копирано в клипборда!",
@@ -616,7 +617,8 @@
     "accountDefaults": "Настройки по подразбиране на акаунта",
     "disableUserSettings": "Деактивиране на страницата с потребителските настройки",
     "userDefaultFieldLockedFromConfig": "Тази настройка по подразбиране е зададена във файла с конфигурацията и не може да бъде променена тук.",
-    "sourceAccessDefaultsHelp": "Прилагат се стандартните права за достъп до файлове за изходните файлове и новите потребители. Опцията „Enforce“ фиксира определена стойност за всички потребители, които не са администратори."
+    "sourceAccessDefaultsHelp": "Прилагат се стандартните права за достъп до файлове за изходните файлове и новите потребители. Опцията „Enforce“ фиксира определена стойност за всички потребители, които не са администратори.",
+    "editAsYaml": "Редактирай като YAML"
   },
   "time": {
     "time": "{prefix}Време{suffix}",
@@ -664,11 +666,13 @@
     "browseIcons": "Разглеждане на икони",
     "pickIcon": "Избор на икона",
     "pickIconDescription": "Търсене и избор на икона от библиотеката с Material Icons",
-    "linksUpdatedSuccess": "Връзките в страничната лента бяха обновени успешно!",
-    "linksUpdateFailed": "Неуспешно обновяване на връзките в страничната лента",
     "sourceNotAccessible": "Този източник вече не е достъпен за вас",
-    "noLinksYet": "Все още няма конфигурирани връзки. Добавете първата си връзка по-долу.",
-    "switchMode": "Смяна на режима на страничната лента"
+    "switchMode": "Смяна на режима на страничната лента",
+    "sidebarLinkDefaults": "Настройки по подразбиране за връзките в страничната лента",
+    "sidebarLinkDefaultsDescription": "Конфигурирайте връзките по подразбиране и задължителните връзки в страничната лента за всички потребители.",
+    "sidebarLinkDefaultsHelp": "Това са връзките по подразбиране, които ще бъдат добавени за всички нови потребители. Ако наложите дадена връзка, тя ще бъде добавена за всички потребители и няма да може да бъде премахната от потребители, които не са администратори.",
+    "sidebarLinkDefaultSourceHelp": "Приложимо е само когато потребителят има достъп до този източник.",
+    "sidebarLinkDefaultsYamlInvalid": "Невалиден YAML. Проверете формата и опитайте отново."
   },
   "api": {
     "title": "API токени",
@@ -1019,8 +1023,7 @@
       "viewActivity": "Преглед на активност",
       "tableColumns": "Колони на таблицата"
     },
-    "toolNotFound": "Инструментът не е намерен",
-    "backToTools": "Обратно към инструментите"
+    "toolNotFound": "Инструментът не е намерен"
   },
   "notifications": {
     "title": "Известия",
