+++ frontend/src/i18n/el.json
@@ -298,7 +298,8 @@
     "accountDefaults": "Προεπιλογές λογαριασμού",
     "disableUserSettings": "Απενεργοποίηση της σελίδας ρυθμίσεων χρήστη",
     "userDefaultFieldLockedFromConfig": "Αυτή η προεπιλογή ορίζεται στο αρχείο ρυθμίσεων και δεν μπορεί να τροποποιηθεί εδώ.",
-    "sourceAccessDefaultsHelp": "Εφαρμόζονται τα προεπιλεγμένα δικαιώματα αρχείων στις πηγές και στους νέους χρήστες. Η επιλογή «Επιβολή» καθορίζει μια συγκεκριμένη τιμή για όλους τους χρήστες που δεν είναι διαχειριστές."
+    "sourceAccessDefaultsHelp": "Εφαρμόζονται τα προεπιλεγμένα δικαιώματα αρχείων στις πηγές και στους νέους χρήστες. Η επιλογή «Επιβολή» καθορίζει μια συγκεκριμένη τιμή για όλους τους χρήστες που δεν είναι διαχειριστές.",
+    "editAsYaml": "Επεξεργασία ως YAML"
   },
   "sidebar": {
     "updateIsAvailable": "Μια ενημέρωση είναι διαθέσιμη!",
@@ -319,10 +320,7 @@
     "linkUrl": "Σύνδεσμος URL",
     "linkUrlPlaceholder": "Εισάγετε τη διεύθυνση URL ή τη διαδρομή (π.χ. /files/path ή https://example.com)",
     "linkIcon": "Εικονίδιο συνδέσμου",
-    "linksUpdatedSuccess": "Οι σύνδεσμοι της πλαϊνής μπάρας ενημερώθηκαν με επιτυχία!",
-    "linksUpdateFailed": "Απέτυχε η ενημέρωση των συνδέσμων της πλευρικής γραμμής",
     "sourceNotAccessible": "Αυτή η πηγή δεν είναι πλέον προσβάσιμη σε εσάς",
-    "noLinksYet": "Δεν έχουν διαμορφωθεί ακόμη σύνδεσμοι. Προσθέστε τον πρώτο σας σύνδεσμο παρακάτω.",
     "customizeShareLinksDescription": "Προσαρμόστε τους συνδέσμους που εμφανίζονται στην πλευρική γραμμή για αυτό το share. Σύρετε και αφήστε για να αναδιατάξετε.",
     "editLink": "Επεξεργασία συνδέσμου",
     "selectShare": "Επιλέξτε Share",
@@ -339,7 +337,12 @@
     "usageTextDisk": "χρήση όπως αναφέρεται από δίσκο/κατάτμηση",
     "dividerNamePlaceholder": "Κείμενο διαχωριστικού (αφήστε το κενό για να εμφανιστεί μια γραμμή)",
     "showIndexedUsage": "Εμφάνιση χρήσης σύμφωνα με το ευρετήριο",
-    "showDiskUsage": "Εμφάνιση χρήσης όπως καταγράφεται ανά δίσκο/διαμέρισμα"
+    "showDiskUsage": "Εμφάνιση χρήσης όπως καταγράφεται ανά δίσκο/διαμέρισμα",
+    "sidebarLinkDefaults": "Προεπιλεγμένες ρυθμίσεις συνδέσμων πλευρικής στήλης",
+    "sidebarLinkDefaultsDescription": "Ρυθμίστε τους προεπιλεγμένους και υποχρεωτικούς συνδέσμους της πλευρικής στήλης για όλους τους χρήστες.",
+    "sidebarLinkDefaultsHelp": "Αυτοί είναι οι προεπιλεγμένοι σύνδεσμοι που θα προστεθούν για όλους τους νέους χρήστες. Εάν ορίσετε έναν σύνδεσμο ως υποχρεωτικό, θα προστεθεί για όλους τους χρήστες και δεν θα μπορεί να αφαιρεθεί από χρήστες που δεν είναι διαχειριστές.",
+    "sidebarLinkDefaultSourceHelp": "Ισχύει μόνο όταν ο χρήστης έχει πρόσβαση σε αυτή την πηγή.",
+    "sidebarLinkDefaultsYamlInvalid": "Το YAML δεν είναι έγκυρο. Ελέγξτε τη μορφή και δοκιμάστε ξανά."
   },
   "share": {
     "notice": "Σημείωση: όταν δημιουργείται μια κοινή χρήση, γίνεται πιο εκτεθειμένη στο κοινό και έχει διαφορετικούς κανόνες πρόσβασης. Εάν η κοινή σας χρήση δεν έχει ρυθμιστεί με ασφάλεια, μπορεί να έχει πρόσβαση ο καθένας.",
@@ -620,7 +623,8 @@
     "limit": "Όριο{suffix}",
     "remove": "Κατάργηση{suffix}",
     "saved": "Αποθηκεύτηκε{suffix}",
-    "removed": "Διαγράφηκε{suffix}"
+    "removed": "Διαγράφηκε{suffix}",
+    "back": "Πίσω{suffix}"
   },
   "access": {
     "rules": "Κανόνες πρόσβασης",
@@ -918,7 +922,6 @@
       "browseFullLibrary": "Αναζήτηση πλήρους βιβλιοθήκης εικονιδίων"
     },
     "toolNotFound": "Το εργαλείο δεν βρέθηκε",
-    "backToTools": "Επιστροφή στα Εργαλεία",
     "fileWatcher": {
       "name": "Παρατηρητής αρχείων",
       "description": "Παρακολούθηση και παρακολούθηση αρχείων σε πραγματικό χρόνο με παρακολούθηση καθυστέρησης",
