+++ frontend/src/i18n/fr.json
@@ -298,7 +298,8 @@
     "accountDefaults": "Paramètres par défaut du compte",
     "disableUserSettings": "Désactiver la page des paramètres utilisateur",
     "userDefaultFieldLockedFromConfig": "Cette valeur par défaut est définie dans le fichier de configuration et ne peut pas être modifiée ici.",
-    "sourceAccessDefaultsHelp": "Droits d'accès par défaut appliqués aux fichiers sources et aux nouveaux utilisateurs. L'option « Enforce » impose une valeur fixe à tous les utilisateurs non administrateurs."
+    "sourceAccessDefaultsHelp": "Droits d'accès par défaut appliqués aux fichiers sources et aux nouveaux utilisateurs. L'option « Enforce » impose une valeur fixe à tous les utilisateurs non administrateurs.",
+    "editAsYaml": "Modifier au format YAML"
   },
   "sidebar": {
     "updateIsAvailable": "Une mise à jour est disponible !",
@@ -319,10 +320,7 @@
     "linkUrl": "Lien URL",
     "linkUrlPlaceholder": "Saisir l'URL ou le chemin d'accès (par exemple, /files/path ou https://example.com)",
     "linkIcon": "Icône de lien",
-    "linksUpdatedSuccess": "Les liens de la barre latérale ont été mis à jour avec succès !",
-    "linksUpdateFailed": "Échec de la mise à jour des liens de la barre latérale",
     "sourceNotAccessible": "Cette source ne vous est plus accessible",
-    "noLinksYet": "Aucun lien n'a encore été configuré. Ajoutez votre premier lien ci-dessous.",
     "customizeShareLinksDescription": "Personnalisez les liens affichés dans la barre latérale pour ce partage. Glissez-déposez pour réorganiser les liens.",
     "editLink": "Modifier le lien",
     "selectShare": "Sélectionner le partage",
@@ -339,7 +337,12 @@
     "usageTextDisk": "utilisation telle que rapportée par disque/partition",
     "dividerNamePlaceholder": "Texte du séparateur (laisser en blanc pour montrer une ligne)",
     "showIndexedUsage": "Afficher l'utilisation telle qu'elle est répertoriée",
-    "showDiskUsage": "Afficher l'utilisation telle qu'elle est indiquée par le disque/la partition"
+    "showDiskUsage": "Afficher l'utilisation telle qu'elle est indiquée par le disque/la partition",
+    "sidebarLinkDefaults": "Paramètres par défaut des liens de la barre latérale",
+    "sidebarLinkDefaultsDescription": "Configurer les liens par défaut et obligatoires de la barre latérale pour tous les utilisateurs.",
+    "sidebarLinkDefaultsHelp": "Il s'agit des liens par défaut qui seront ajoutés pour tous les nouveaux utilisateurs. Si vous rendez un lien obligatoire, celui-ci sera ajouté pour tous les utilisateurs et ne pourra pas être supprimé par les utilisateurs n'ayant pas les droits d'administrateur.",
+    "sidebarLinkDefaultSourceHelp": "S'applique uniquement lorsque l'utilisateur a accès à cette source.",
+    "sidebarLinkDefaultsYamlInvalid": "YAML non valide. Vérifiez le format et réessayez."
   },
   "share": {
     "notice": "Remarque : lorsqu'un partage est créé, il est davantage exposé au public et les règles d'accès sont différentes. Si votre partage n'est pas configuré de manière sécurisée, n'importe qui peut y accéder.",
@@ -620,7 +623,8 @@
     "limit": "Limite{suffix}",
     "remove": "Supprimer{suffix}",
     "saved": "Enregistré{suffix}",
-    "removed": "Supprimé{suffix}"
+    "removed": "Supprimé{suffix}",
+    "back": "Retour{suffix}"
   },
   "access": {
     "rules": "Règles d'accès",
@@ -918,7 +922,6 @@
       "browseFullLibrary": "Parcourir la bibliothèque complète d'icônes"
     },
     "toolNotFound": "Outil non trouvé",
-    "backToTools": "Retour aux outils",
     "fileWatcher": {
       "name": "Observateur de fichiers",
       "description": "Surveillance et suivi des fichiers en temps réel grâce au contrôle de la latence",
