+++ frontend/src/i18n/he.json
@@ -298,7 +298,8 @@
     "accountDefaults": "הגדרות ברירת מחדל של החשבון",
     "disableUserSettings": "השבת את דף הגדרות המשתמש",
     "userDefaultFieldLockedFromConfig": "הגדרה ברירת המחדל הזו נקבעת בקובץ התצורה ואינה ניתנת לשינוי כאן.",
-    "sourceAccessDefaultsHelp": "הרשאות הקובץ המוגדרות כברירת מחדל מוחלות על מקורות ועל משתמשים חדשים. האפשרות \"אכוף\" נועלת ערך זה עבור כל המשתמשים שאינם מנהלים."
+    "sourceAccessDefaultsHelp": "הרשאות הקובץ המוגדרות כברירת מחדל מוחלות על מקורות ועל משתמשים חדשים. האפשרות \"אכוף\" נועלת ערך זה עבור כל המשתמשים שאינם מנהלים.",
+    "editAsYaml": "ערוך כ-YAML"
   },
   "sidebar": {
     "updateIsAvailable": "עדכון זמין!",
@@ -319,10 +320,7 @@
     "linkUrl": "כתובת URL של הקישור",
     "linkUrlPlaceholder": "הזן כתובת URL או נתיב (לדוגמה, /files/path או https://example.com)",
     "linkIcon": "סמל קישור",
-    "linksUpdatedSuccess": "קישורי הסרגל הצדדי עודכנו בהצלחה!",
-    "linksUpdateFailed": "לא הצלחנו לעדכן את הקישורים בסרגל הצדדי",
     "sourceNotAccessible": "מקור זה אינו נגיש לך עוד",
-    "noLinksYet": "עדיין לא הוגדרו קישורים. הוסף את הקישור הראשון שלך למטה.",
     "customizeShareLinksDescription": "התאם אישית את הקישורים המוצגים בסרגל הצדדי עבור שיתוף זה. גרור ושחרר כדי לשנות את הסדר.",
     "editLink": "ערוך קישור",
     "selectShare": "בחר שיתוף",
@@ -339,7 +337,12 @@
     "usageTextDisk": "שימוש כפי שדווח על ידי דיסק/מחיצה",
     "dividerNamePlaceholder": "טקסט מפריד (השאר ריק כדי להציג קו)",
     "showIndexedUsage": "הצג את השימוש כפי שמופיע באינדקס",
-    "showDiskUsage": "הצג את השימוש כפי שדווח על ידי הדיסק/המחיצה"
+    "showDiskUsage": "הצג את השימוש כפי שדווח על ידי הדיסק/המחיצה",
+    "sidebarLinkDefaults": "הגדרות ברירת מחדל לקישורים בסרגל הצדדי",
+    "sidebarLinkDefaultsDescription": "הגדר קישורים ברירת מחדל וקישורים מחייבים בסרגל הצדדי עבור כל המשתמשים.",
+    "sidebarLinkDefaultsHelp": "אלה הם הקישורים המוגדרים כברירת מחדל שייתווספו לכל המשתמשים החדשים. אם תגדיר קישור כמחויב, הוא יתווסף לכל המשתמשים ולא ניתן יהיה להסיר אותו על ידי משתמשים שאינם מנהלים.",
+    "sidebarLinkDefaultSourceHelp": "חל רק כאשר למשתמש יש גישה למקור זה.",
+    "sidebarLinkDefaultsYamlInvalid": "YAML לא תקין. בדוק את הפורמט ונסה שוב."
   },
   "share": {
     "notice": "הערה: כאשר נוצר שיתוף, הוא הופך לחשוף יותר לציבור וחל עליו כללים שונים לגבי גישה. אם השיתוף שלך אינו מוגדר בצורה מאובטחת, כל אחד יכול לגשת אליו.",
@@ -620,7 +623,8 @@
     "limit": "גבול{suffix}",
     "remove": "הסר{suffix}",
     "saved": "נשמר{suffix}",
-    "removed": "הוסר{suffix}"
+    "removed": "הוסר{suffix}",
+    "back": "חזרה{suffix}"
   },
   "access": {
     "rules": "כללי גישה",
@@ -918,7 +922,6 @@
       "browseFullLibrary": "עיין בספריית הסמלים המלאה"
     },
     "toolNotFound": "הכלי לא נמצא",
-    "backToTools": "חזרה לכלים",
     "fileWatcher": {
       "name": "צופה קבצים",
       "description": "צפה בקבצים ועקוב אחריהם בזמן אמת עם ניטור חביון",
