+++ frontend/src/i18n/ro.json
@@ -298,7 +298,8 @@
     "accountDefaults": "Setări implicite ale contului",
     "disableUserSettings": "Dezactivare pagină setări utilizator",
     "userDefaultFieldLockedFromConfig": "Această setare implicită este definită în fișierul de configurare și nu poate fi modificată aici.",
-    "sourceAccessDefaultsHelp": "Permisiunile implicite aplicate fișierelor sursă și utilizatorilor noi. Opțiunea „Enforce” blochează o valoare pentru toți utilizatorii care nu sunt administratori."
+    "sourceAccessDefaultsHelp": "Permisiunile implicite aplicate fișierelor sursă și utilizatorilor noi. Opțiunea „Enforce” blochează o valoare pentru toți utilizatorii care nu sunt administratori.",
+    "editAsYaml": "Editează ca YAML"
   },
   "sidebar": {
     "updateIsAvailable": "O actualizare este disponibilă!",
@@ -319,10 +320,7 @@
     "linkUrl": "Link URL",
     "linkUrlPlaceholder": "Introduceți adresa URL sau calea (de exemplu, /files/path sau https://example.com)",
     "linkIcon": "Pictograma de legătură",
-    "linksUpdatedSuccess": "Legăturile din bara laterală au fost actualizate cu succes!",
-    "linksUpdateFailed": "A eșuat actualizarea linkurilor din bara laterală",
     "sourceNotAccessible": "Această sursă nu vă mai este accesibilă",
-    "noLinksYet": "Nicio legătură configurată încă. Adăugați primul dvs. link mai jos.",
     "customizeShareLinksDescription": "Personalizați linkurile afișate în bara laterală pentru această partajare. Glisați și fixați pentru a le reordona.",
     "editLink": "Editare link",
     "selectShare": "Selectați Partajare",
@@ -339,7 +337,12 @@
     "usageTextDisk": "utilizare raportată de disc/partiție",
     "dividerNamePlaceholder": "Textul separatorului (lăsați gol pentru a afișa o linie)",
     "showIndexedUsage": "Afișează utilizarea conform indexului",
-    "showDiskUsage": "Afișează utilizarea conform datelor raportate de disc/partiție"
+    "showDiskUsage": "Afișează utilizarea conform datelor raportate de disc/partiție",
+    "sidebarLinkDefaults": "Setări implicite pentru linkurile din bara laterală",
+    "sidebarLinkDefaultsDescription": "Configurați linkurile implicite și obligatorii din bara laterală pentru toți utilizatorii.",
+    "sidebarLinkDefaultsHelp": "Acestea sunt linkurile implicite care vor fi adăugate pentru toți utilizatorii noi. Dacă impui un link, acesta va fi adăugat pentru toți utilizatorii și nu poate fi eliminat de către utilizatorii care nu sunt administratori.",
+    "sidebarLinkDefaultSourceHelp": "Se aplică numai în cazul în care utilizatorul are acces la această sursă.",
+    "sidebarLinkDefaultsYamlInvalid": "YAML nevalid. Verificați formatul și încercați din nou."
   },
   "share": {
     "notice": "Notă: atunci când este creată o partajare, aceasta devine mai expusă publicului și are reguli diferite de acces. Dacă partajul dvs. nu este configurat în siguranță, acesta poate fi accesat de oricine.",
@@ -620,7 +623,8 @@
     "limit": "Limită{suffix}",
     "remove": "Eliminare{suffix}",
     "saved": "Salvat{suffix}",
-    "removed": "Eliminat{suffix}"
+    "removed": "Eliminat{suffix}",
+    "back": "Înapoi{suffix}"
   },
   "access": {
     "rules": "Reguli de acces",
@@ -918,7 +922,6 @@
       "browseFullLibrary": "Răsfoiți biblioteca completă de pictograme"
     },
     "toolNotFound": "Instrumentul nu a fost găsit",
-    "backToTools": "Înapoi la Instrumente",
     "fileWatcher": {
       "name": "Observator de fișiere",
       "description": "Urmăriți și urmăriți fișierele în timp real cu monitorizarea latenței",
