+++ frontend/src/i18n/ru.json
@@ -298,7 +298,8 @@
     "accountDefaults": "Настройки по умолчанию учетной записи",
     "disableUserSettings": "Отключить страницу настроек пользователя",
     "userDefaultFieldLockedFromConfig": "Это значение по умолчанию задано в файле конфигурации и не может быть изменено здесь.",
-    "sourceAccessDefaultsHelp": "К исходным файлам и новым пользователям применяются права доступа по умолчанию. Параметр «Enforce» фиксирует значение для всех пользователей, не являющихся администраторами."
+    "sourceAccessDefaultsHelp": "К исходным файлам и новым пользователям применяются права доступа по умолчанию. Параметр «Enforce» фиксирует значение для всех пользователей, не являющихся администраторами.",
+    "editAsYaml": "Редактировать в формате YAML"
   },
   "sidebar": {
     "updateIsAvailable": "Обновление доступно!",
@@ -319,10 +320,7 @@
     "linkUrl": "Ссылка URL",
     "linkUrlPlaceholder": "Введите URL-адрес или путь (например, /files/path или https://example.com).",
     "linkIcon": "Значок ссылки",
-    "linksUpdatedSuccess": "Ссылки в сайдбаре успешно обновлены!",
-    "linksUpdateFailed": "Не удалось обновить ссылки боковой панели",
     "sourceNotAccessible": "Этот источник больше недоступен для вас",
-    "noLinksYet": "Пока нет настроенных ссылок. Добавьте свою первую ссылку ниже.",
     "customizeShareLinksDescription": "Настройте ссылки, отображаемые в боковой панели для этого ресурса. Перетащите ссылки, чтобы изменить их порядок.",
     "editLink": "Редактировать ссылку",
     "selectShare": "Выберите долю",
@@ -339,7 +337,12 @@
     "usageTextDisk": "использование по данным диска/раздела",
     "dividerNamePlaceholder": "Текст разделителя (оставьте пустым, чтобы показать строку)",
     "showIndexedUsage": "Показать использование в соответствии с индексом",
-    "showDiskUsage": "Показать объем использования по данным диска/раздела"
+    "showDiskUsage": "Показать объем использования по данным диска/раздела",
+    "sidebarLinkDefaults": "Настройки по умолчанию для ссылок в боковой панели",
+    "sidebarLinkDefaultsDescription": "Настройте ссылки в боковой панели, которые будут отображаться по умолчанию и обязательны для всех пользователей.",
+    "sidebarLinkDefaultsHelp": "Это ссылки по умолчанию, которые будут добавлены для всех новых пользователей. Если вы сделаете ссылку обязательной, она будет добавлена для всех пользователей, и пользователи, не являющиеся администраторами, не смогут её удалить.",
+    "sidebarLinkDefaultSourceHelp": "Применимо только в том случае, если у пользователя есть доступ к данному источнику.",
+    "sidebarLinkDefaultsYamlInvalid": "Неверный YAML. Проверьте формат и попробуйте ещё раз."
   },
   "share": {
     "notice": "Примечание: когда создается общий ресурс, он становится более открытым для публики и имеет другие правила доступа. Если ваш ресурс не настроен безопасно, доступ к нему может получить любой желающий.",
@@ -620,7 +623,8 @@
     "limit": "Предел{suffix}",
     "remove": "Удалить{suffix}",
     "saved": "Сохранено{suffix}",
-    "removed": "Удалено{suffix}"
+    "removed": "Удалено{suffix}",
+    "back": "Назад{suffix}"
   },
   "access": {
     "rules": "Правила доступа",
@@ -918,7 +922,6 @@
       "browseFullLibrary": "Просмотр полной библиотеки иконок"
     },
     "toolNotFound": "Инструмент не найден",
-    "backToTools": "Назад к инструментам",
     "fileWatcher": {
       "name": "Наблюдатель за файлами",
       "description": "Наблюдение за файлами в режиме реального времени с контролем задержки",
