+++ frontend/src/i18n/ua.json
@@ -298,7 +298,8 @@
     "accountDefaults": "Стандартні налаштування облікового запису",
     "disableUserSettings": "Вимкнути сторінку налаштувань користувача",
     "userDefaultFieldLockedFromConfig": "Це значення за замовчуванням задається у файлі конфігурації, і його неможливо змінити тут.",
-    "sourceAccessDefaultsHelp": "Застосовано стандартні права доступу до файлів для джерел та нових користувачів. Параметр «Enforce» фіксує значення для всіх користувачів, які не є адміністраторами."
+    "sourceAccessDefaultsHelp": "Застосовано стандартні права доступу до файлів для джерел та нових користувачів. Параметр «Enforce» фіксує значення для всіх користувачів, які не є адміністраторами.",
+    "editAsYaml": "Редагувати у форматі YAML"
   },
   "sidebar": {
     "updateIsAvailable": "Доступне оновлення!",
@@ -319,10 +320,7 @@
     "linkUrl": "URL-адреса посилання",
     "linkUrlPlaceholder": "Введіть URL або шлях (наприклад, /files/path або https://example.com)",
     "linkIcon": "Значок посилання",
-    "linksUpdatedSuccess": "Посилання на бічну панель успішно оновлено!",
-    "linksUpdateFailed": "Не вдалося оновити посилання бічної панелі",
     "sourceNotAccessible": "Це джерело більше не доступне для вас",
-    "noLinksYet": "Ще немає налаштованих посилань. Додайте своє перше посилання нижче.",
     "customizeShareLinksDescription": "Налаштуйте посилання, показані на бічній панелі для цього ресурсу. Перетягніть, щоб змінити порядок.",
     "editLink": "Редагувати посилання",
     "selectShare": "Виберіть \"Поділитися",
@@ -339,7 +337,12 @@
     "usageTextDisk": "використання за диском/розділом",
     "dividerNamePlaceholder": "Текст роздільника (залиште порожнім, щоб показати лінію)",
     "showIndexedUsage": "Показати використання відповідно до індексу",
-    "showDiskUsage": "Показати використання за даними диска/розділу"
+    "showDiskUsage": "Показати використання за даними диска/розділу",
+    "sidebarLinkDefaults": "Стандартні налаштування посилань у бічній панелі",
+    "sidebarLinkDefaultsDescription": "Налаштувати посилання на бічній панелі, які відображатимуться за замовчуванням та обов’язково для всіх користувачів.",
+    "sidebarLinkDefaultsHelp": "Це посилання, які додаватимуться за замовчуванням для всіх нових користувачів. Якщо ви зробите посилання обов’язковим, воно додаватиметься для всіх користувачів, і користувачі, які не є адміністраторами, не зможуть його видалити.",
+    "sidebarLinkDefaultSourceHelp": "Застосовується лише в тому випадку, якщо користувач має доступ до цього джерела.",
+    "sidebarLinkDefaultsYamlInvalid": "Неправильний формат YAML. Перевірте формат і спробуйте ще раз."
   },
   "share": {
     "notice": "Примітка: після створення спільного ресурсу він стає більш відкритим для публіки і має інші правила доступу. Якщо ваш спільний ресурс не налаштовано безпечно, до нього може отримати доступ будь-хто.",
@@ -718,7 +721,8 @@
     "limit": "Обмеження{suffix}",
     "remove": "Видалити{suffix}",
     "saved": "Збережено{suffix}",
-    "removed": "Видалено{suffix}"
+    "removed": "Видалено{suffix}",
+    "back": "Назад{suffix}"
   },
   "access": {
     "rules": "Правила доступу",
@@ -918,7 +922,6 @@
       "browseFullLibrary": "Переглянути повну бібліотеку іконок"
     },
     "toolNotFound": "Інструмент не знайдено",
-    "backToTools": "Повернутися до Інструментів",
     "fileWatcher": {
       "name": "Переглядач файлів",
       "description": "Перегляд і відстеження файлів у реальному часі з моніторингом затримок",
