+++ frontend/src/i18n/zh-tw.json
@@ -298,7 +298,8 @@
     "accountDefaults": "帳戶預設值",
     "disableUserSettings": "停用使用者設定頁面",
     "userDefaultFieldLockedFromConfig": "此預設值是在設定檔中設定的，無法在此處進行變更。",
-    "sourceAccessDefaultsHelp": "套用預設檔案權限至原始檔及新使用者。強制鎖定所有非管理員使用者的權限值。"
+    "sourceAccessDefaultsHelp": "套用預設檔案權限至原始檔及新使用者。強制鎖定所有非管理員使用者的權限值。",
+    "editAsYaml": "以 YAML 格式編輯"
   },
   "sidebar": {
     "updateIsAvailable": "有可用的更新！",
@@ -319,10 +320,7 @@
     "linkUrl": "連結網址",
     "linkUrlPlaceholder": "輸入 URL 或路徑 (例如 /files/path 或 https://example.com)",
     "linkIcon": "連結圖示",
-    "linksUpdatedSuccess": "側邊欄連結更新成功！",
-    "linksUpdateFailed": "更新側邊欄連結失敗",
     "sourceNotAccessible": "您無法再存取此來源",
-    "noLinksYet": "尚未設定任何連結。請在下方新增第一個連結。",
     "customizeShareLinksDescription": "自訂此分享側邊欄顯示的連結。可拖放重新排序。",
     "editLink": "編輯連結",
     "selectShare": "選擇分享",
@@ -339,7 +337,12 @@
     "usageTextDisk": "依磁碟/磁碟分割回報的使用量",
     "dividerNamePlaceholder": "分隔線文字（留空則只顯示一條線）",
     "showIndexedUsage": "顯示索引中的用法",
-    "showDiskUsage": "顯示根據磁碟／分割區所報告的使用情況"
+    "showDiskUsage": "顯示根據磁碟／分割區所報告的使用情況",
+    "sidebarLinkDefaults": "側邊欄連結的預設值",
+    "sidebarLinkDefaultsDescription": "為所有使用者設定預設及強制顯示的側邊欄連結。",
+    "sidebarLinkDefaultsHelp": "這些是將為所有新使用者預設新增的連結。若您強制啟用某個連結，該連結將會新增至所有使用者的帳戶中，且非管理員無法移除。",
+    "sidebarLinkDefaultSourceHelp": "僅當使用者具有此來源的存取權限時，才適用此規定。",
+    "sidebarLinkDefaultsYamlInvalid": "YAML 格式無效。請檢查格式並重新嘗試。"
   },
   "share": {
     "notice": "注意：建立分享後，內容會更容易對外公開，且存取規則也不同。如果未妥善設定安全性，任何人都可能存取。",
@@ -718,7 +721,8 @@
     "limit": "上限{suffix}",
     "remove": "移除{suffix}",
     "saved": "已儲存{suffix}",
-    "removed": "已移除{suffix}"
+    "removed": "已移除{suffix}",
+    "back": "返回{suffix}"
   },
   "access": {
     "rules": "存取規則",
@@ -918,7 +922,6 @@
       "browseFullLibrary": "瀏覽完整圖示庫"
     },
     "toolNotFound": "找不到工具",
-    "backToTools": "回到工具",
     "fileWatcher": {
       "name": "檔案監看器",
       "description": "即時監看並追蹤檔案內容，同時監測延遲",
